:round_pushpin: sys_draw2d

+++ src/math/vec2.rs
@@ -1,7 +1,7 @@
 #[derive(Clone, Copy)]
 pub struct Vec2 {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Vec2 {
@@ -9,7 +9,7 @@ impl Vec2 {
         Vec2::new(0.0, 0.0)
     }
 
-    pub fn new(x: f64, y: f64, ) -> Vec2 {
+    pub fn new(x: f32, y: f32, ) -> Vec2 {
         Vec2 { x, y }
     }
 }
